Fix frontend dependencies
