+++ grassroots-backend/src/grassroots-shared/util/CreateDTOBase.ts
@@ -38,10 +38,11 @@ export interface FetchResponse<T, E> {
 export function createDTOBase<TBrand extends string>(brand: TBrand) {
   abstract class Branded {
     // We need different names for our branding types
-    // between entities or DTOs, or we get some type collicions.
+    // between entities or DTOs, or we get some type collisions.
     readonly __DTOBrand!: `${TBrand}DTO`;
     // Used for CASL to identify object types.
     readonly __caslSubjectType: TBrand = brand;
+    static readonly __caslSubjectTypeStatic: TBrand = brand;
 
     static from<T extends Branded>(
       // The this parameter must be named "this", and is magically populated with the class constructor.
@@ -67,5 +68,6 @@ export function createDTOBase<TBrand extends string>(brand: TBrand) {
       );
     }
   }
+
   return Branded;
 }
