+++ grassroots-backend/tsconfig.json
@@ -11,7 +11,12 @@
     "sourceMap": true,
     "outDir": "./dist",
     "incremental": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "paths": {
+      "@shared/CASLSubjects": [
+        "./src/grassroots-shared/CASLSubjects/CASLSubjects.backend.ts"
+      ]
+    }
   },
   "exclude": ["src/testing/typetests/**"]
 }
